Correct README example
Fix broken example in `README.md` and also fix bad code formatting that
broke the build. Also, in the example, renamed variables `request` and
`response` to `req` and `resp`, respectively, to avoid confusion with
modules of the same name. While the duplicative names do not cause any
compilation or runtime failures, they do cause confusion for the reader.